LayoutTests/pointer-lock/pointer-lock-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../http/tests/resources/js-test-pre.js"></script>
</head>
<body>
<script>
    description("Basic API existence test for Pointer Lock.")

    shouldBeDefined("document.onwebkitpointerlockchange");
    shouldBeDefined("document.onwebkitpointerlockerror");
    shouldBeDefined("document.webkitExitPointerLock");
    shouldEvaluateTo("document.webkitPointerLockElement", null);
    element = document.createElement();
    shouldBeDefined("element.webkitRequestPointerLock");
</script>
<script src="../http/tests/resources/js-test-post.js"></script>
</body>
</html>
