LayoutTests/pointer-lock/pointer-lock-api-expected.txt

Basic API existence test for Pointer Lock.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.onwebkitpointerlockchange is defined.
PASS document.onwebkitpointerlockerror is defined.
PASS document.webkitExitPointerLock is defined.
PASS document.webkitPointerLockElement is null
PASS element.webkitRequestPointerLock is defined.
PASS successfullyParsed is true

TEST COMPLETE
